(svn r5999) -Feature: change the original date format to a 32 bits format based at the year 0. The game date subsystem now allows someone to start in the year 0 and continue up to the year 5 000 000. However, you currently cannot build anything before 1920 as there is no newgrf support for dates before 1920 or after 2090 yet.

+++ openttd.h
@@ -54,11 +54,11 @@ typedef byte WindowClass;
 
 enum {
 	INVALID_YEAR = -1,
-	INVALID_DATE = (uint16)-1,
+	INVALID_DATE = -1,
 };
 
-typedef int16  Year;
-typedef uint16 Date;
+typedef int32 Year;
+typedef int32 Date;
 
 
 enum GameModes {
